+++ app/src/main/java/ru/karapetiandav/yamblzproject/ui/fragments/AboutFragment.java
@@ -1,4 +1,4 @@
-package ru.karapetiandav.yamblzproject.fragments;
+package ru.karapetiandav.yamblzproject.ui.fragments;
 
 
 import android.os.Bundle;
@@ -9,14 +9,14 @@ import android.view.ViewGroup;
 
 import ru.karapetiandav.yamblzproject.R;
 
-public class SettingsFragment extends Fragment {
+public class AboutFragment extends Fragment {
 
-    public SettingsFragment() {
+    public AboutFragment() {
         // Required empty public constructor
     }
 
-    public static SettingsFragment newInstance() {
-        SettingsFragment fragment = new SettingsFragment();
+    public static AboutFragment newInstance() {
+        AboutFragment fragment = new AboutFragment();
 
         return fragment;
     }
@@ -25,7 +25,7 @@ public class SettingsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        return inflater.inflate(R.layout.fragment_settings, container, false);
+        return inflater.inflate(R.layout.fragment_about, container, false);
     }
 
 }
