+++ app/src/main/res/layout/fragment_about.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ru.karapetiandav.yamblzproject.fragments.AboutFragment">
+    tools:context=".ui.fragments.AboutFragment">
 
     <TextView
         android:layout_width="match_parent"
