+++ app/src/main/res/layout/fragment_settings.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ru.karapetiandav.yamblzproject.fragments.SettingsFragment">
+    tools:context=".ui.fragments.SettingsFragment">
 
     <TextView
         android:layout_width="wrap_content"
