+++ app/src/main/res/layout/fragment_weather.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ru.karapetiandav.yamblzproject.fragments.WeatherFragment">
+    tools:context=".ui.fragments.WeatherFragment">
 
     <ImageView
         android:id="@+id/weather_placeholder"
